+++ app/(root)/(tabs)/chat.tsx
@@ -8,7 +8,7 @@ import ISConnectedCard from "@/components/ISConnectedCard";
 const Chat = () => {
   const { state } = useNetworkCheck();
   return (
-    <SafeAreaView className="flex-1 bg-white p-5">
+    <SafeAreaView className="flex-1 bg-white p-5 dark:bg-custom-dark ">
       {!state.isConnected && <ISConnectedCard customClass="!ml-5" />}
       <ScrollView contentContainerStyle={{ flexGrow: 1 }}>
         <Text className="text-2xl font-JakartaBold">Chat</Text>